# swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "API Adoção de Animais",
    "version": "1.0.0",
    "description": "Backend para site de adoção de animais baseado em metamodelo (archetypes/entities/attributes)."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/v1",
      "description": "Desenvolvimento local"
    }
  ],
  "tags": [
    { "name": "Auth" },
    { "name": "Archetypes" },
    { "name": "Entities" },
    { "name": "AdoptionRequests" }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Registro de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Login de usuário",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login bem sucedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthTokens"
                }
              }
            }
          },
          "401": {
            "description": "Credenciais inválidas"
          }
        }
      }
    },
    "/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "Usuário autenticado",
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Dados do usuário logado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserPublic"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado"
          }
        }
      }
    },
    "/archetypes": {
      "get": {
        "tags": ["Archetypes"],
        "summary": "Lista arquetipos",
        "responses": {
          "200": {
            "description": "Lista de arquetipos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Archetype"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Archetypes"],
        "summary": "Cria novo arquetipo",
        "security": [
          { "bearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchetypeCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Archetype"
                }
              }
            }
          }
        }
      }
    },
    "/archetypes/{id}": {
      "get": {
        "tags": ["Archetypes"],
        "summary": "Detalhes de arquetipo",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Arquetipo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Archetype"
                }
              }
            }
          },
          "404": { "description": "Não encontrado" }
        }
      },
      "patch": {
        "tags": ["Archetypes"],
        "summary": "Atualiza arquetipo",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArchetypeUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Archetype"
                }
              }
            }
          }
        }
      }
    },
    "/entities": {
      "get": {
        "tags": ["Entities"],
        "summary": "Lista entidades",
        "parameters": [
          {
            "in": "query",
            "name": "archetypeSlug",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "status",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "page",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": { "type": "integer", "default": 20 }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista paginada de entidades",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedEntities"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Entities"],
        "summary": "Cria nova entidade",
        "security": [
          { "bearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetail"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{id}": {
      "get": {
        "tags": ["Entities"],
        "summary": "Detalhes de entidade",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Entidade com atributos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetail"
                }
              }
            }
          },
          "404": { "description": "Não encontrado" }
        }
      },
      "patch": {
        "tags": ["Entities"],
        "summary": "Atualiza entidade",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EntityUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EntityDetail"
                }
              }
            }
          }
        }
      }
    },
    "/entities/{id}/attributes": {
      "get": {
        "tags": ["Entities"],
        "summary": "Lista valores de atributos da entidade",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de atributos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityAttributeValue"
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": ["Entities"],
        "summary": "Atualiza valores de atributos",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true,
                "description": "Mapa attributeKey -> valor"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/EntityAttributeValue"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/adoption-requests": {
      "post": {
        "tags": ["AdoptionRequests"],
        "summary": "Cria pedido de adoção",
        "security": [
          { "bearerAuth": [] }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdoptionRequestCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Pedido criado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdoptionRequest"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["AdoptionRequests"],
        "summary": "Lista pedidos (admin)",
        "security": [
          { "bearerAuth": [] }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": { "type": "string" }
          },
          {
            "in": "query",
            "name": "entityId",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdoptionRequest"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/adoption-requests/my": {
      "get": {
        "tags": ["AdoptionRequests"],
        "summary": "Lista pedidos do usuário logado",
        "security": [
          { "bearerAuth": [] }
        ],
        "responses": {
          "200": {
            "description": "Lista de pedidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdoptionRequest"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "RegisterRequest": {
        "type": "object",
        "required": ["name", "email", "password"],
        "properties": {
          "name": { "type": "string" },
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" }
        }
      },
      "LoginRequest": {
        "type": "object",
        "required": ["email", "password"],
        "properties": {
          "email": { "type": "string", "format": "email" },
          "password": { "type": "string", "format": "password" }
        }
      },
      "AuthTokens": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" }
        }
      },
      "UserPublic": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "emailVerified": { "type": "boolean" }
        }
      },
      "Archetype": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" },
          "isActive": { "type": "boolean" }
        }
      },
      "ArchetypeCreate": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": {
          "name": { "type": "string" },
          "slug": { "type": "string" },
          "description": { "type": "string" }
        }
      },
      "ArchetypeUpdate": {
        "type": "object",
        "properties": {
          "name": { "type": "string" },
          "description": { "type": "string" },
          "isActive": { "type": "boolean" }
        }
      },
      "EntityBase": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "uuid": { "type": "string", "format": "uuid" },
          "archetypeId": { "type": "integer" },
          "title": { "type": "string" },
          "status": { "type": "string" }
        }
      },
      "EntityAttributeValue": {
        "type": "object",
        "properties": {
          "attributeId": { "type": "integer" },
          "attributeKey": { "type": "string" },
          "valueText": { "type": "string" },
          "valueNumber": { "type": "number" },
          "valueBool": { "type": "boolean" },
          "valueDate": { "type": "string", "format": "date" },
          "valueJson": { "type": "object" }
        }
      },
      "EntityDetail": {
        "allOf": [
          { "$ref": "#/components/schemas/EntityBase" },
          {
            "type": "object",
            "properties": {
              "attributes": {
                "type": "array",
                "items": { "$ref": "#/components/schemas/EntityAttributeValue" }
              }
            }
          }
        ]
      },
      "EntityCreate": {
        "type": "object",
        "required": ["archetypeId", "title"],
        "properties": {
          "archetypeId": { "type": "integer" },
          "title": { "type": "string" },
          "status": { "type": "string", "default": "draft" },
          "attributes": {
            "type": "object",
            "additionalProperties": true
          }
        }
      },
      "EntityUpdate": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "status": { "type": "string" }
        }
      },
      "PaginatedEntities": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EntityDetail" }
          },
          "page": { "type": "integer" },
          "pageSize": { "type": "integer" },
          "total": { "type": "integer" }
        }
      },
      "AdoptionRequest": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "entityId": { "type": "integer" },
          "requesterUserId": { "type": "integer" },
          "status": {
            "type": "string",
            "enum": ["PENDING", "ACCEPTED", "REJECTED", "CANCELLED"]
          },
          "formDataJson": { "type": "object" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
      "AdoptionRequestCreate": {
        "type": "object",
        "required": ["entityId", "formData"],
        "properties": {
          "entityId": { "type": "integer" },
          "formData": {
            "type": "object",
            "description": "Dados do formulário de adoção"
          }
        }
      }
    }
  }
}